create salaryable module and create class helper file
