+++ class_helpers.rb
@@ -1,0 +1,3 @@
+require_relative 'person'
+require_relative 'phase_settable'
+require_relative 'salary_setter'
